requirements.txt
# ── Core ──────────────────────────────────────────────
pandas>=1.5,<3
numpy>=1.23,<2
scikit-learn>=1.2,<2
scipy>=1.10

# ── Visualization ─────────────────────────────────────
matplotlib>=3.6
seaborn>=0.12

# ── Mining / Association Rules ────────────────────────
mlxtend>=0.22

# ── Gradient Boosting ─────────────────────────────────
xgboost>=1.7
lightgbm>=3.3

# ── Time Series ───────────────────────────────────────
statsmodels>=0.13

# ── Config & IO ───────────────────────────────────────
pyyaml>=6.0
pyarrow>=10.0          # parquet read/write

# ── Notebook automation (optional) ────────────────────
papermill>=2.4
jupyter>=1.0
ipykernel>=6.0
